refactor: use dynamic value eval for objects in other renderers too
