+++ frontend/src/components/StudioComponent.vue
@@ -171,30 +171,6 @@ const evaluationContext = computed(() => {
 	}
 })
 
-const evaluateDynamicValues = (value: any): any => {
-	if (typeof value === "string") {
-		if (isDynamicValue(value)) {
-			const evaluated = codeStore.getDynamicValue(value, evaluationContext.value)
-			return evaluated
-		}
-		return value
-	}
-
-	if (Array.isArray(value)) {
-		return value.map((item) => evaluateDynamicValues(item))
-	}
-
-	if (value !== null && typeof value === "object") {
-		const result: Record<string, any> = {}
-		for (const [key, val] of Object.entries(value)) {
-			result[key] = evaluateDynamicValues(val)
-		}
-		return result
-	}
-
-	return value
-}
-
 const getComponentProps = () => {
 	if (!props.block || props.block.isRoot()) return []
 
@@ -202,7 +178,7 @@ const getComponentProps = () => {
 	delete propValues.modelValue
 
 	Object.entries(propValues).forEach(([propName, config]) => {
-		propValues[propName] = evaluateDynamicValues(config)
+		propValues[propName] = codeStore.evaluateDynamicValues(config, evaluationContext.value)
 	})
 	return propValues
 }
@@ -220,8 +196,7 @@ const componentRef = ref<ComponentPublicInstance | null>(null)
 // visibility
 const showComponent = computed(() => {
 	if (props.block.visibilityCondition) {
-		const value = codeStore.getDynamicValue(props.block.visibilityCondition, evaluationContext.value)
-		return value
+		return codeStore.getDynamicValue(props.block.visibilityCondition, evaluationContext.value)
 	}
 	return true
 })
